pyproject.toml
[project]
name = "toon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=3.1.0",
    "python-toon>=0.1.3",
    "tiktoken>=0.12.0",
    "zon-format>=1.2.0",
]
[project.optional-dependencies]
research = [
    "dspy-ai>=3.0.4",
    "tiktoken>=0.4.0",
    "python-toon>=0.1.3",
    "pydantic>=2.0.0",
    "psutil>=7.1.3",
]

[project.urls]
Repository = "https://github.com/marcusjihansson/memory_opt"
